AbstractEnumConfigConverter moved
