+++ hedera-node/hedera-config/src/main/java/com/hedera/node/config/converter/ProfileConverter.java
@@ -18,6 +18,7 @@ package com.hedera.node.config.converter;
 
 import com.hedera.node.config.types.Profile;
 import com.swirlds.config.api.converter.ConfigConverter;
+import com.swirlds.config.extensions.converters.AbstractEnumConfigConverter;
 
 public class ProfileConverter extends AbstractEnumConfigConverter<Profile> implements ConfigConverter<Profile> {
     @Override
