+++ platform-sdk/swirlds-config-extensions/build.gradle.kts
@@ -23,4 +23,6 @@ testModuleInfo {
     requires("com.swirlds.common")
     requires("com.swirlds.test.framework")
     requires("org.junit.jupiter.api")
+    requires("org.junit.jupiter.params")
+    requires("org.assertj.core")
 }
