+++ platform-sdk/swirlds-config-extensions/src/main/java/module-info.java
@@ -3,6 +3,7 @@ module com.swirlds.config.extensions {
     exports com.swirlds.config.extensions.reflection;
     exports com.swirlds.config.extensions.sources;
     exports com.swirlds.config.extensions.validators;
+    exports com.swirlds.config.extensions.converters;
 
     requires transitive com.swirlds.config.api;
     requires com.swirlds.base;
